requirements.txt
asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
cachetools==5.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.52.0
django-cors-headers==4.1.0
django-daraja==1.3.0
django-phonenumber-field==7.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
paypalrestsdk==1.13.1
phonenumbers==8.13.22
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.2.1
pyrsistent==0.19.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.34.0
whitenoise==6.9.0
